requirements.txt
torch
torchtext
pytext-nlp